--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,4 @@
+dist
 out
 node_modules
 .vscode-test/