--- a/src/configure/clients/devOps/azureDevOpsClient.ts
+++ b/src/configure/clients/devOps/azureDevOpsClient.ts
@@ -31,7 +31,7 @@ export class AzureDevOpsClient {
     }
 
     public async createOrganization(organizationName: string): Promise<any> {
-        return this.sendRequest(<RequestPrepareOptions>{
+        return this.sendRequest({
             url: "https://app.vsaex.visualstudio.com/_apis/HostAcquisition/collections",
             headers: {
                 "Content-Type": "application/json"
@@ -50,7 +50,7 @@ export class AzureDevOpsClient {
     public async createProject(organizationName: string, projectName: string): Promise<any> {
         let collectionUrl = `https://dev.azure.com/${organizationName}`;
 
-        return this.sendRequest(<RequestPrepareOptions>{
+        return this.sendRequest({
             url: `${collectionUrl}/_apis/projects`,
             headers: {
                 "Content-Type": "application/json"
@@ -84,7 +84,7 @@ export class AzureDevOpsClient {
         if (!this.listOrgPromise || forceRefresh) {
             this.listOrgPromise = this.getUserData()
                 .then((connectionData) => {
-                    return this.sendRequest(<RequestPrepareOptions>{
+                    return this.sendRequest({
                         url: "https://app.vssps.visualstudio.com/_apis/accounts",
                         headers: {
                             "Content-Type": "application/json"
@@ -111,7 +111,7 @@ export class AzureDevOpsClient {
 
     public async listProjects(organizationName: string): Promise<Array<DevOpsProject>> {
         let url = `${AzureDevOpsBaseUrl}/${organizationName}/_apis/projects`;
-        let response = await this.sendRequest(<RequestPrepareOptions>{
+        let response = await this.sendRequest({
             url: url,
             headers: {
                 "Content-Type": "application/json"
@@ -137,7 +137,7 @@ export class AzureDevOpsClient {
     public async getRepository(organizationName: string, projectName: string, repositoryName: string): Promise<any> {
         let url = `${AzureDevOpsBaseUrl}/${organizationName}/${projectName}/_apis/git/repositories/${repositoryName}`;
 
-        return this.sendRequest(<RequestPrepareOptions>{
+        return this.sendRequest({
             url: url,
             headers: {
                 "Content-Type": "application/json",
@@ -154,7 +154,7 @@ export class AzureDevOpsClient {
     public async createBuildDefinition(organizationName: string, buildDefinition: BuildDefinition): Promise<any> {
         let url = `${AzureDevOpsBaseUrl}/${organizationName}/${buildDefinition.project.id}/_apis/build/definitions`;
 
-        return this.sendRequest(<RequestPrepareOptions>{
+        return this.sendRequest({
             url: url,
             method: "POST",
             headers: {
@@ -169,7 +169,7 @@ export class AzureDevOpsClient {
     public async queueBuild(organizationName: string, build: Build): Promise<any> {
         let url = `${AzureDevOpsBaseUrl}/${organizationName}/${build.project.id}/_apis/build/builds`;
 
-        return this.sendRequest(<RequestPrepareOptions>{
+        return this.sendRequest({
             url: url,
             method: "POST",
             headers: {
@@ -194,7 +194,7 @@ export class AzureDevOpsClient {
         else {
             let url = `https://app.vsaex.visualstudio.com/_apis/HostAcquisition/NameAvailability/${organizationName}`;
 
-            this.sendRequest(<RequestPrepareOptions>{
+            this.sendRequest({
                 url: url,
                 headers: {
                     "Content-Type": "application/json",
@@ -220,7 +220,7 @@ export class AzureDevOpsClient {
     public async getProjectIdFromName(organizationName: string, projectName: string): Promise<string> {
         let url = `${AzureDevOpsBaseUrl}/${organizationName}/_apis/projects/${projectName}`;
 
-        return this.sendRequest(<RequestPrepareOptions>{
+        return this.sendRequest({
             url: url,
             method: "GET",
             headers: {
@@ -275,7 +275,7 @@ export class AzureDevOpsClient {
     }
 
     private getConnectionData(): Promise<any> {
-        return this.sendRequest(<RequestPrepareOptions>{
+        return this.sendRequest({
             url: "https://app.vssps.visualstudio.com/_apis/connectiondata",
             headers: {
                 "Content-Type": "application/json"
@@ -287,7 +287,7 @@ export class AzureDevOpsClient {
     }
 
     private createUserProfile(): Promise<any> {
-        return this.sendRequest(<RequestPrepareOptions>{
+        return this.sendRequest({
             url: "https://app.vssps.visualstudio.com/_apis/_AzureProfile/CreateProfile",
             headers: {
                 "Content-Type": "application/json"
@@ -321,7 +321,7 @@ export class AzureDevOpsClient {
     }
 
     private async getOperationResult(operationUrl: string): Promise<any> {
-        return this.sendRequest(<RequestPrepareOptions>{
+        return this.sendRequest({
             url: operationUrl,
             queryParameters: {
                 "api-version": "5.0"
@@ -335,7 +335,7 @@ export class AzureDevOpsClient {
     public getAgentQueues(organizationName: string, projectName: string): Promise<Array<any>> {
         let url = `${AzureDevOpsBaseUrl}/${organizationName}/${projectName}/_apis/distributedtask/queues`;
 
-        return this.sendRequest(<RequestPrepareOptions>{
+        return this.sendRequest({
             url: url,
             method: "GET",
             headers: {