--- a/src/configure/clients/devOps/serviceConnectionClient.ts
+++ b/src/configure/clients/devOps/serviceConnectionClient.ts
@@ -1,7 +1,7 @@
 import { AadApplication } from '../../model/models';
 import { AzureDevOpsBaseUrl } from "../../resources/constants";
 import { AzureDevOpsClient } from './azureDevOpsClient';
-import { UrlBasedRequestPrepareOptions } from 'ms-rest';
+import { RequestPrepareOptions } from '@azure/ms-rest-js';
 
 export class ServiceConnectionClient {
     private azureDevOpsClient: AzureDevOpsClient;
@@ -17,7 +17,7 @@ export class ServiceConnectionClient {
     public async createGitHubServiceConnection(endpointName: string, gitHubPat: string): Promise<any> {
         let url = `${AzureDevOpsBaseUrl}/${this.organizationName}/${this.projectName}/_apis/serviceendpoint/endpoints`;
 
-        return this.azureDevOpsClient.sendRequest(<UrlBasedRequestPrepareOptions>
+        return this.azureDevOpsClient.sendRequest(<RequestPrepareOptions>
             {
                 url: url,
                 headers: {
@@ -49,7 +49,7 @@ export class ServiceConnectionClient {
     public async createAzureServiceConnection(endpointName: string, tenantId: string, subscriptionId: string, scope: string, aadApp: AadApplication): Promise<any> {
         let url = `${AzureDevOpsBaseUrl}/${this.organizationName}/${this.projectName}/_apis/serviceendpoint/endpoints`;
 
-        return this.azureDevOpsClient.sendRequest(<UrlBasedRequestPrepareOptions>{
+        return this.azureDevOpsClient.sendRequest(<RequestPrepareOptions>{
             url: url,
             headers: {
                 "Content-Type": "application/json",
@@ -89,13 +89,13 @@ export class ServiceConnectionClient {
     public async getEndpointStatus(endpointId: string): Promise<any> {
         let url = `${AzureDevOpsBaseUrl}/${this.organizationName}/${this.projectName}/_apis/serviceendpoint/endpoints/${endpointId}`;
 
-        return this.azureDevOpsClient.sendRequest(<UrlBasedRequestPrepareOptions>{
+        return this.azureDevOpsClient.sendRequest(<RequestPrepareOptions>{
             url: url,
             headers: {
                 "Content-Type": "application/json",
                 "Accept": "application/json;api-version=5.1-preview.2;excludeUrls=true"
             },
-            method: "Get",
+            method: "GET",
             deserializationMapper: null,
             serializationMapper: null
         });
@@ -104,7 +104,7 @@ export class ServiceConnectionClient {
     public async authorizeEndpointForAllPipelines(endpointId: string): Promise<any> {
         let url = `${AzureDevOpsBaseUrl}/${this.organizationName}/${this.projectName}/_apis/pipelines/pipelinePermissions/endpoint/${endpointId}`;
 
-        return this.azureDevOpsClient.sendRequest(<UrlBasedRequestPrepareOptions>{
+        return this.azureDevOpsClient.sendRequest(<RequestPrepareOptions>{
             url: url,
             headers: {
                 "Content-Type": "application/json",