… .vscode/launch.json
… package.json
… src/configure/clients/restClient.ts
… src/configure/configure.ts
… src/configure/helper/azureSessionHelper.ts
… src/configure/helper/graphHelper.ts
… src/configure/helper/templateHelper.ts
… src/configure/model/models.ts
… src/configure/resources/messages.ts
… src/extension.ts
… tsconfig.json
… webpack.config.js