Introducing **faster activation times** and **smaller extension downloads**!

Activation time with pipeline configuration enabled:
* Before: 9736ms (not a typo, not sure what happened here), 1492ms, 1366ms
* After: 215ms, 252ms, 292ms

Extension size:
* Before: 9.86MB
* After: ~~2.59MB~~ 1.17MB

So we have _very_ snappy activation time now and a ~~75%~~ 90% smaller download size :).

I need to do further manual testing for configure pipeline on ADO repositories to make sure nothing regressed due to the amount of package upgrades that needed to happen in that area (I've verified that I get a valid access token back but didn't want to test actual pipeline creation on my Microsoft account, hehe). Once I'm reasonably certain that it's still working, I'll merge.

Definitively fixes #215. I see a comment there about brittle webpacking with the language server so I'll make sure to follow up on that.
Fixes #381

TODO:
* [x] ~~Fix azureDevOpsClient authentication~~