@50Wliu 2021-02-05
> Fix azureDevOpsClient authentication

Sooooo this never worked......you can't use a token for management.core.azure.net on Azure DevOps, which is what the old code was trying to do.
I am still unfamiliar with MSAL but preliminary investigation shows that fixing this is way out of the scope of this PR.